2009 and 2013 is NIOSII
The first folder, from the previous 5 years of this course, are when we used the ARM processor and ARM Assembly language, so the questions from those years, together with the solutions are in an assembly language we did not use this year - Nios II Assembly.  They are included so you can see the types of questions that were asked, and you may wish to try and solve these questions as they relate to Nios II Assembly.  Some questions are related to the C programming language, and they would still be relevant.  In addition, the order in which the course was taught changed in 2022, so this folder provides a selection of questions from the final exam that may be helpful from 2019-2021.

The second folder contains questions from the midterm tests in this course between 2009 and 2013, when the course also used the Nios II processor and assembly language.  However, there were different emphasis in the course in those years.  Professor Moshovos has marked, in red, questions that he thinks are not relevant to this years' course.  

There may be typographic errors in these older documents.